Addes collisons with lava
Added a collison with the lava that sets the player health to 0 and resets the current scene
