+++ Shadow Keep/Assets/Player/Prefabs/Player.prefab
@@ -15,6 +15,7 @@ GameObject:
   - component: {fileID: 3104680088112795285}
   - component: {fileID: 2719564349655288611}
   - component: {fileID: 7314502854372070245}
+  - component: {fileID: 918559429178042349}
   m_Layer: 0
   m_Name: Player
   m_TagString: Untagged
@@ -56,6 +57,9 @@ MonoBehaviour:
   myRidgidBody: {fileID: 2574388804705511358}
   mySprite: {fileID: 3104680088112795285}
   animator: {fileID: 2719564349655288611}
+  playerCollider: {fileID: 5492349592508965365}
+  groundCollider: {fileID: 7314502854372070245}
+  playerInformationScript: {fileID: 918559429178042349}
 --- !u!50 &2574388804705511358
 Rigidbody2D:
   serializedVersion: 5
@@ -243,3 +247,17 @@ CapsuleCollider2D:
   m_Offset: {x: 0.041989606, y: -0.24120104}
   m_Size: {x: 0.36951703, y: 0.07433607}
   m_Direction: 1
+--- !u!114 &918559429178042349
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4079736338124651580}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c847bf20bb4b8402da01ad2d7f058f7d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  health: 100
+  power: 100
