+++ Shadow Keep/Assets/PlayerInformationScript.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c847bf20bb4b8402da01ad2d7f058f7d
